+++ src/main/java/net/ldm/mo_food/block/GarlicCrop.java
@@ -1,0 +1,17 @@
+package net.ldm.mo_food.block;
+
+import net.ldm.mo_food.init.MoFood;
+import net.minecraft.block.PotatoesBlock;
+import net.minecraft.item.ItemConvertible;
+
+public class GarlicCrop extends PotatoesBlock {
+
+    public GarlicCrop( Settings settings ) {
+        super(settings);
+    }
+
+    @Override
+    protected ItemConvertible getSeedsItem() {
+        return MoFood.GARLIC;
+    }
+}