+++ src/main/java/net/ldm/mo_food/block/TomatoesBlock.java
@@ -1,4 +1,3 @@
-/*
 package net.ldm.mo_food.block;
 
 import net.fabricmc.api.EnvType;
@@ -121,4 +120,3 @@ public class TomatoesBlock extends PlantBlock implements Fertilizable {
         LARGE_SHAPE = Block.createCuboidShape(1.0D, 0.0D, 1.0D, 15.0D, 16.0D, 15.0D);
     }
 }
-*/