+++ src/main/java/net/ldm/mo_food/init/ClientInit.java
@@ -10,7 +10,7 @@ public class ClientInit implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         BlockRenderLayerMap.INSTANCE.putBlock(SWEET_POTATOES, RenderLayer.getCutout());
-        //BlockRenderLayerMap.INSTANCE.putBlock(TOMATOES, RenderLayer.getCutout());
+        BlockRenderLayerMap.INSTANCE.putBlock(TOMATOES, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(LETTUCE_CROP, RenderLayer.getCutout());
         BlockRenderLayerMap.INSTANCE.putBlock(GARLIC_CROP, RenderLayer.getCutout());
         //ScreenRegistry.register(MoFood.COOKTOP_SCREEN_HANDLER, CooktopScreen::new);